Replace Network+Storage preferences with Downloads
