+++ app/src/main/java/de/danoeh/antennapod/fragment/preferences/DownloadsPreferencesFragment.java
@@ -8,25 +8,29 @@ import androidx.preference.PreferenceManager;
 import de.danoeh.antennapod.R;
 import de.danoeh.antennapod.activity.PreferenceActivity;
 import de.danoeh.antennapod.core.util.download.FeedUpdateManager;
+import de.danoeh.antennapod.dialog.ChooseDataFolderDialog;
 import de.danoeh.antennapod.dialog.ProxyDialog;
 import de.danoeh.antennapod.storage.preferences.UserPreferences;
 
+import java.io.File;
 
-public class NetworkPreferencesFragment extends PreferenceFragmentCompat
+
+public class DownloadsPreferencesFragment extends PreferenceFragmentCompat
         implements SharedPreferences.OnSharedPreferenceChangeListener {
     private static final String PREF_SCREEN_AUTODL = "prefAutoDownloadSettings";
     private static final String PREF_PROXY = "prefProxy";
+    private static final String PREF_CHOOSE_DATA_DIR = "prefChooseDataDir";
 
     @Override
     public void onCreatePreferences(Bundle savedInstanceState, String rootKey) {
-        addPreferencesFromResource(R.xml.preferences_network);
+        addPreferencesFromResource(R.xml.preferences_downloads);
         setupNetworkScreen();
     }
 
     @Override
     public void onStart() {
         super.onStart();
-        ((PreferenceActivity) getActivity()).getSupportActionBar().setTitle(R.string.network_pref);
+        ((PreferenceActivity) getActivity()).getSupportActionBar().setTitle(R.string.downloads_pref);
         PreferenceManager.getDefaultSharedPreferences(getContext()).registerOnSharedPreferenceChangeListener(this);
     }
 
@@ -40,6 +44,7 @@ public class NetworkPreferencesFragment extends PreferenceFragmentCompat
     public void onResume() {
         super.onResume();
         setParallelDownloadsText(UserPreferences.getParallelDownloads());
+        setDataFolderText();
     }
 
     private void setupNetworkScreen() {
@@ -59,12 +64,26 @@ public class NetworkPreferencesFragment extends PreferenceFragmentCompat
             dialog.show();
             return true;
         });
+        findPreference(PREF_CHOOSE_DATA_DIR).setOnPreferenceClickListener(preference -> {
+            ChooseDataFolderDialog.showDialog(getContext(), path -> {
+                UserPreferences.setDataFolder(path);
+                setDataFolderText();
+            });
+            return true;
+        });
     }
 
     private void setParallelDownloadsText(int downloads) {
         final Resources res = getActivity().getResources();
         String s = res.getString(R.string.parallel_downloads, downloads);
         findPreference(UserPreferences.PREF_PARALLEL_DOWNLOADS).setSummary(s);
+    }
+
+    private void setDataFolderText() {
+        File f = UserPreferences.getDataFolder(null);
+        if (f != null) {
+            findPreference(PREF_CHOOSE_DATA_DIR).setSummary(f.getAbsolutePath());
+        }
     }
 
     @Override
