+++ app/src/main/java/de/danoeh/antennapod/fragment/preferences/MainPreferencesFragment.java
@@ -30,9 +30,9 @@ public class MainPreferencesFragment extends PreferenceFragmentCompat {
 
     private static final String PREF_SCREEN_USER_INTERFACE = "prefScreenInterface";
     private static final String PREF_SCREEN_PLAYBACK = "prefScreenPlayback";
-    private static final String PREF_SCREEN_NETWORK = "prefScreenNetwork";
+    private static final String PREF_SCREEN_DOWNLOADS = "prefScreenDownloads";
+    private static final String PREF_SCREEN_IMPORT_EXPORT = "prefScreenImportExport";
     private static final String PREF_SCREEN_SYNCHRONIZATION = "prefScreenSynchronization";
-    private static final String PREF_SCREEN_STORAGE = "prefScreenStorage";
     private static final String PREF_DOCUMENTATION = "prefDocumentation";
     private static final String PREF_VIEW_FORUM = "prefViewForum";
     private static final String PREF_SEND_BUG_REPORT = "prefSendBugReport";
@@ -90,16 +90,16 @@ public class MainPreferencesFragment extends PreferenceFragmentCompat {
             ((PreferenceActivity) getActivity()).openScreen(R.xml.preferences_playback);
             return true;
         });
-        findPreference(PREF_SCREEN_NETWORK).setOnPreferenceClickListener(preference -> {
-            ((PreferenceActivity) getActivity()).openScreen(R.xml.preferences_network);
+        findPreference(PREF_SCREEN_DOWNLOADS).setOnPreferenceClickListener(preference -> {
+            ((PreferenceActivity) getActivity()).openScreen(R.xml.preferences_downloads);
             return true;
         });
         findPreference(PREF_SCREEN_SYNCHRONIZATION).setOnPreferenceClickListener(preference -> {
             ((PreferenceActivity) getActivity()).openScreen(R.xml.preferences_synchronization);
             return true;
         });
-        findPreference(PREF_SCREEN_STORAGE).setOnPreferenceClickListener(preference -> {
-            ((PreferenceActivity) getActivity()).openScreen(R.xml.preferences_storage);
+        findPreference(PREF_SCREEN_IMPORT_EXPORT).setOnPreferenceClickListener(preference -> {
+            ((PreferenceActivity) getActivity()).openScreen(R.xml.preferences_import_export);
             return true;
         });
         findPreference(PREF_NOTIFICATION).setOnPreferenceClickListener(preference -> {
@@ -166,15 +166,12 @@ public class MainPreferencesFragment extends PreferenceFragmentCompat {
                 .addBreadcrumb(PreferenceActivity.getTitleOfPage(R.xml.preferences_user_interface));
         config.index(R.xml.preferences_playback)
                 .addBreadcrumb(PreferenceActivity.getTitleOfPage(R.xml.preferences_playback));
-        config.index(R.xml.preferences_network)
-                .addBreadcrumb(PreferenceActivity.getTitleOfPage(R.xml.preferences_network));
-        config.index(R.xml.preferences_storage)
-                .addBreadcrumb(PreferenceActivity.getTitleOfPage(R.xml.preferences_storage));
+        config.index(R.xml.preferences_downloads)
+                .addBreadcrumb(PreferenceActivity.getTitleOfPage(R.xml.preferences_downloads));
         config.index(R.xml.preferences_import_export)
-                .addBreadcrumb(PreferenceActivity.getTitleOfPage(R.xml.preferences_storage))
                 .addBreadcrumb(PreferenceActivity.getTitleOfPage(R.xml.preferences_import_export));
         config.index(R.xml.preferences_autodownload)
-                .addBreadcrumb(PreferenceActivity.getTitleOfPage(R.xml.preferences_network))
+                .addBreadcrumb(PreferenceActivity.getTitleOfPage(R.xml.preferences_downloads))
                 .addBreadcrumb(R.string.automation)
                 .addBreadcrumb(PreferenceActivity.getTitleOfPage(R.xml.preferences_autodownload));
         config.index(R.xml.preferences_synchronization)
