+++ app/src/main/res/xml/preferences.xml
@@ -22,10 +22,10 @@
             android:icon="@drawable/ic_play_24dp" />
 
     <Preference
-            android:key="prefScreenNetwork"
-            android:title="@string/network_pref"
-            android:summary="@string/network_pref_sum"
-            android:icon="@drawable/ic_network" />
+            android:key="prefScreenDownloads"
+            android:title="@string/downloads_pref"
+            android:summary="@string/downloads_pref_sum"
+            android:icon="@drawable/ic_download_black" />
 
     <Preference
             android:key="prefScreenSynchronization"
@@ -34,9 +34,9 @@
             android:icon="@drawable/ic_cloud" />
 
     <Preference
-            android:key="prefScreenStorage"
-            android:title="@string/storage_pref"
-            android:summary="@string/storage_sum"
+            android:key="prefScreenImportExport"
+            android:title="@string/import_export_pref"
+            android:summary="@string/import_export_summary"
             android:icon="@drawable/ic_storage" />
 
     <Preference
