--- app/src/main/res/xml/preferences_network.xml
@@ -1,47 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<PreferenceScreen
-        xmlns:android="http://schemas.android.com/apk/res/android"
-        xmlns:numberpicker="http://schemas.android.com/apk/de.danoeh.antennapod"
-        xmlns:search="http://schemas.android.com/apk/com.bytehamster.lib.preferencesearch">
-    <PreferenceCategory android:title="@string/automation">
-        <de.danoeh.antennapod.preferences.MaterialListPreference
-                android:entryValues="@array/feed_refresh_interval_values"
-                android:entries="@array/feed_refresh_interval_entries"
-                android:key="prefAutoUpdateIntervall"
-                android:title="@string/feed_refresh_title"
-                android:summary="@string/feed_refresh_sum"
-                android:defaultValue="12"/>
-        <Preference
-                android:summary="@string/pref_automatic_download_sum"
-                android:key="prefAutoDownloadSettings"
-                android:title="@string/pref_automatic_download_title"
-                search:ignore="true" />
-        <de.danoeh.antennapod.preferences.MaterialListPreference
-                android:entryValues="@array/globalNewEpisodesActionValues"
-                android:entries="@array/globalNewEpisodesActionItems"
-                android:key="prefNewEpisodesAction"
-                android:title="@string/pref_new_episodes_action_title"
-                android:summary="@string/pref_new_episodes_action_sum"
-                android:defaultValue="1"/>
-    </PreferenceCategory>
-
-    <PreferenceCategory android:title="@string/download_pref_details">
-        <MultiSelectListPreference
-                android:defaultValue="@array/mobile_update_default_value"
-                android:entries="@array/mobile_update_entries"
-                android:entryValues="@array/mobile_update_values"
-                android:key="prefMobileUpdateTypes"
-                android:summary="@string/pref_mobileUpdate_sum"
-                android:title="@string/pref_mobileUpdate_title"/>
-        <de.danoeh.antennapod.preferences.NumberPickerPreference
-                android:defaultValue="4"
-                numberpicker:minValue="1"
-                numberpicker:maxValue="50"
-                android:key="prefParallelDownloads"
-                android:title="@string/pref_parallel_downloads_title"/>
-        <Preference
-                android:key="prefProxy"
-                android:summary="@string/pref_proxy_sum"
-                android:title="@string/pref_proxy_title"/>
-    </PreferenceCategory>
-</PreferenceScreen>
