--- app/src/main/res/xml/preferences_storage.xml
@@ -1,32 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<PreferenceScreen
-        xmlns:android="http://schemas.android.com/apk/res/android"
-        xmlns:search="http://schemas.android.com/apk/com.bytehamster.lib.preferencesearch">
-
-    <Preference
-            android:title="@string/choose_data_directory"
-            android:key="prefChooseDataDir"/>
-    <SwitchPreferenceCompat
-            android:defaultValue="false"
-            android:enabled="true"
-            android:key="prefAutoDelete"
-            android:summary="@string/pref_auto_delete_sum"
-            android:title="@string/pref_auto_delete_title"/>
-    <SwitchPreferenceCompat
-            android:defaultValue="true"
-            android:enabled="true"
-            android:key="prefFavoriteKeepsEpisode"
-            android:summary="@string/pref_favorite_keeps_episodes_sum"
-            android:title="@string/pref_favorite_keeps_episodes_title"/>
-    <SwitchPreferenceCompat
-            android:defaultValue="false"
-            android:enabled="true"
-            android:key="prefDeleteRemovesFromQueue"
-            android:summary="@string/pref_delete_removes_from_queue_sum"
-            android:title="@string/pref_delete_removes_from_queue_title"/>
-    <Preference
-            android:title="@string/import_export_pref"
-            android:summary="@string/import_export_summary"
-            android:key="prefImportExport"
-            search:ignore="true"/>
-</PreferenceScreen>
